requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-model-utils==4.5.1
python-dotenv==1.0.0
psycopg2-binary==2.9.9
flake8==5.0.4